// src/enums/HttpStausCode.ts
enum HttpStatusCode {

    CONTINUE = 100,

    SWITCHING_PROTOCOLS = 101,

    PROCESSING = 102,

    OK = 200,

    CREATED = 201,

    ACCEPTED = 202,

    NON_AUTHORITATIVE_INFORMATION = 203,

    NO_CONTENT = 204,

    MOVED_PERMANENTLY = 301,

    FOUND = 302,

    BAD_REQUEST = 400,

    UNAUTHORIZED = 401,

    FORBIDDEN = 403,

    NOT_FOUND = 404,

    METHOD_NOT_ALLOWED = 405,

    INTERNAL_SERVER_ERROR = 500,

    NOT_IMPLEMENTED = 501,

    BAD_GATEWAY = 502,

}

export default HttpStatusCode;